requirements.txt
APScheduler==3.6.3
astroid==2.6.2
autopep8==1.5.7
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
pycodestyle==2.7.0
pylint==2.9.3
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
wrapt==1.12.1
